Factor out some methods to a GitCli utility class
